Feat: 대댓글 조회 쿼리 작성(#1)

+++ src/main/java/com/cdd/recipeservice/recipemodule/comment/dto/cond/RecipeRootCommentFindCond.java
@@ -2,7 +2,7 @@ package com.cdd.recipeservice.recipemodule.comment.dto.cond;
 
 public record RecipeRootCommentFindCond(
 	PagingCond pagingCond,
-	long recipeId
+	int recipeId
 ) {
 
 	public static RecipeRootCommentFindCond searchRootComment(
